pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "advanced-analysis-for-salla"
version = "1.0.0"
description = "Local, offline analytics tool for Salla XLSX exports"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.29.0",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.1.9",
    "scikit-learn>=1.3.2",
    "scipy>=1.11.4",
    "plotly>=5.18.0",
    "babel>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.7",
    "black>=23.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]
performance = [
    "polars>=0.19.19",
]
fuzzy = [
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.23.0",
]
rtl = [
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "plotly.*",
    "sklearn.*",
    "scipy.*",
    "arabic_reshaper.*",
    "bidi.*",
    "fuzzywuzzy.*",
    "polars.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"